clean functions that are already in oggm main, add test
