+++ __init__.py
@@ -1,2 +1,2 @@
 # This is a dummy init.
-__version__ = '0.0.1'
+__version__ = '1.0.1'
