+++ utils.py
@@ -6,6 +6,7 @@ import numpy as np
 import copy
 from functools import (wraps)
 
+
 def nicenumber(number, binsize, lower=False):
     """Returns the next higher or lower "nice number", given by binsize.
     Examples:
@@ -154,6 +155,87 @@ class SuperclassMeta(type):
                     pass
         return cls
 
+##################################################################
+# Light version of Centerline class from OGGM
+class Centerline(object, metaclass=SuperclassMeta):
+    """Geometry (line and widths) and flow rooting properties, but no thickness
+    """
+
+    def __init__(self, line, dx=None, surface_h=None, orig_head=None,
+                  rgi_id=None, map_dx=None):
+        """ Initialize a Centerline
+        Parameters
+        ----------
+        line : :py:class:`shapely.geometry.LineString`
+            The geometrically calculated centerline
+        dx : float
+            Grid spacing of the initialised flowline in pixel coordinates
+        surface_h :  :py:class:`numpy.ndarray`
+            elevation [m] of the points on ``line``
+        orig_head : :py:class:`shapely.geometry.Point`
+            geometric point of the lines head
+        rgi_id : str
+            The glacier's RGI identifier
+        map_dx : float
+            the map's grid resolution. Centerline.dx_meter = dx * map_dx
+        """
+
+        self.line = None  # Shapely LineString
+        self.head = None  # Shapely Point
+        self.tail = None  # Shapely Point
+        self.dis_on_line = None
+        self.nx = None
+        if line is not None:
+            self.set_line(line)  # Init all previous properties
+        else:
+            self.nx = len(surface_h)
+            self.dis_on_line = np.arange(self.nx) * dx
+
+        self.order = None  # Hydrological flow level (~ Strahler number)
+
+        # These are computed at run time by compute_centerlines
+        self.flows_to = None  # pointer to a Centerline object (when available)
+        self.flows_to_point = None  # point of the junction in flows_to
+        self.inflows = []  # list of Centerline instances (when available)
+        self.inflow_points = []  # junction points
+
+        # Optional attrs
+        self.dx = dx  # dx in pixels (assumes the line is on constant dx
+        self.map_dx = map_dx  # the pixel spacing
+        try:
+            self.dx_meter = self.dx * self.map_dx
+        except TypeError:
+            # For backwards compatibility we allow this for now
+            self.dx_meter = None
+        self._surface_h = surface_h
+        self._widths = None
+        self.is_rectangular = None
+        self.is_trapezoid = None
+        self.orig_head = orig_head  # Useful for debugging and for filtering
+        self.geometrical_widths = None  # these are kept for plotting and such
+        self.apparent_mb = None  # Apparent MB, NOT weighted by width.
+        self.mu_star = None  # the mu* associated with the apparent mb
+        self.mu_star_is_valid = False  # if mu* leeds to good flux, keep it
+        self.flux = None  # Flux (kg m-2)
+        self.flux_needs_correction = False  # whether this branch was baaad
+        self.rgi_id = rgi_id  # Useful if line is used with another glacier
+
+
+    def set_line(self, line):
+        """Update the Shapely LineString coordinate.
+        Parameters
+        ----------
+        line : :py:class`shapely.geometry.LineString`
+        """
+
+        self.nx = len(line.coords)
+        self.line = line
+        dis = [line.project(shpg.Point(co)) for co in line.coords]
+        self.dis_on_line = np.array(dis)
+        xx, yy = line.xy
+        self.head = shpg.Point(xx[0], yy[0])
+        self.tail = shpg.Point(xx[-1], yy[-1])
+
 
 def line_interpol(line, dx):
     """Interpolates a shapely LineString to a regularly spaced one.
@@ -239,3 +321,23 @@ def cls_to_geoline(cls):
         liness.append(cl.line)
     lines = liness
     return lines
+
+def geoline_to_cls(lines):
+    """
+    list of shapely.geometry.lines to be converted to Centerline list 
+
+    Parameters
+    ----------
+    lines : list of shapely.geometry.lines
+        list of shapely.geometry.lines
+
+    Returns
+    -------
+    list of centerlines instances.
+
+    """
+    clss = []
+    for li in lines:
+        clss.append(Centerline(li))
+    cls = clss
+    return cls
